refactor: utilize nested wrapping (#9196)
Refs #8020
