+++ src/core/plugins/oas3/components/auth/http-auth.jsx
