+++ src/core/plugins/oas3/wrap-components/index.js
@@ -1,5 +1,5 @@
 import Markdown from "./markdown"
-import AuthItem from "./auth-item"
+import AuthItem from "./auth/auth-item"
 import OnlineValidatorBadge from "./online-validator-badge"
 import Model from "./model"
 import JsonSchema_string from "./json-schema-string"
