+++ src/core/plugins/oas31/index.js
@@ -9,15 +9,15 @@ import JsonSchemaDialect from "./components/json-schema-dialect"
 import VersionPragmaFilter from "./components/version-pragma-filter"
 import Model from "./components/model/model"
 import Models from "./components/models/models"
-import MutualTLSAuth from "./components/mutual-tls-auth"
-import Auths from "./components/auths"
+import MutualTLSAuth from "./components/auth/mutual-tls-auth"
+import Auths from "./components/auth/auths"
 import LicenseWrapper from "./wrap-components/license"
 import ContactWrapper from "./wrap-components/contact"
 import InfoWrapper from "./wrap-components/info"
 import ModelWrapper from "./wrap-components/model"
 import ModelsWrapper from "./wrap-components/models"
 import VersionPragmaFilterWrapper from "./wrap-components/version-pragma-filter"
-import AuthItemWrapper from "./wrap-components/auth-item"
+import AuthItemWrapper from "./wrap-components/auth/auth-item"
 import AuthsWrapper from "./wrap-components/auths"
 import {
   isOAS31 as isOAS31Fn,
